+++ Backends/Graphics5/G5onG4/Sources/Kore/Shader5Impl.c
@@ -1,10 +1,8 @@
 #include "pch.h"
 
-#include <Kinc/Graphics5/Shader.h>
-#include <Kinc/Math/Core.h>
+#include <kinc/graphics4/shader.h>
+#include <kinc/graphics5/shader.h>
 
-#include <Kinc/Graphics4/Shader.h>
-
-void kinc_g5_shader_init(kinc_g5_shader_t *shader, void *source, int length, kinc_g5_shader_type_t type) {
+void kinc_g5_shader_init(kinc_g5_shader_t *shader, void *source, size_t length, kinc_g5_shader_type_t type) {
 	kinc_g4_shader_init(&shader->impl.shader, source, length, (kinc_g4_shader_type_t)type);
 }
